refactor(sdk): rename _oss_client.py to oss_client.py
OssClient 是 Sandbox 公开组合的成员（self._oss），底层模块没必要 _ 私有前缀。
